Rework #chapter to support nested divisions.
Formerly, the chapter method was stupid: it used a passed block as an
alternate way of providing the title and filename attributes; if you
needed to calculate those parameters dynamically (for example,
fetching the data for the file from someplace as part of the script)
you would do that in the block and return the title and filename,
rather than passing them in as parameters.

How long have I been using Ruby again? If you need to do that, write a
method that does that and then pass them into chapter.

So now that's the way it works.  I needed the block for other things.
Like supporting hierarchical document structures.

Now chapter has several aliases (section, part, appendix) as well as
an underlying div method that you can use to build arbitrary division
types (such as "book"). Any block passed to those methods receives the
division object as a parameter, and you can call any of those methods
within the block, allowing for (as just one example) chapters nested
within higher-level "parts".